requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0